usersFile and User class.
I’ve added a new method that writes all the users to a seperate file.
Both writeToXml() and writeUsersToXml() are called whenever an issue is
added or updated. If a user is added, only the writeUsersToFile()
method is called. I’ve also added a new class called User, but haven’t
done anything with it yet. Will change the users list to contain User
objects instead of just strings at a later time. The priority search is
now also changed to the 5 different Strings as I forgot this last
commit. Btw, IntelliJ rules.