Added fields in Issues
Added the fields status, createdBy and updatedBy in the Issues class.
I’ve only added the status field to the issues constructor so far and
added this to the lists and JTable. However, when it comes to the other
to fields and some of the other tasks in the assignment I feel we need
to do a small redesign of the gui. We need some more space to add
additional components such as redirecting output stream and maybe
showing the actually issue text in a bigger text field for instance. I
would prefer doing fullscreen mode to begin with. That would create
some space.