Added better layout and CustomOutputStream
Created 2 new classes. The CustomOutputStream class redirects the
System.out.prints() to the textArea in the TablePanel class. The table
panel class is the class that holds the JTable and the customOutStream.
The next thing on the list is to add another panel on the .EAST side of
the spine panel in gui. This new panel will present more specific info
about the current selected row, such as last updatedBy and createdBy
and so on(not decided how to structure yet).