Removed Issue column and added a new Panel
I’ve removed the issue column containing the issue info and made it
easier to read an issue. The new DetailsPanel will now display the
issue of the current selected row in the JTable. Next on the list is
the get the Created by and updated by to work and write them to file.