Changed created to Date
I’ve changed the created field in the Issues class to now be a Date
instead of a String. Even though it is stored as a date, it will be
presented as a string using the method dateToString() in issueTable.
The date strings in the xml file are converted to dates using the
stringToDate() method in issueTable(). Also some minor bug fixes. I
know this was off topic for this branch, but making tests is so boring.
Will do it now however…